refactor: Update verification methods to use decimal instead of double and add Converter utility class

+++ Core/Core.Tax/Dsl/Commands/ReturnsTaxRate.cs
@@ -29,7 +29,7 @@ public class ReturnsTaxRate : BaseTaxCommand<VoidValue, VoidVerification>
         return this;
     }
 
-    public ReturnsTaxRate TaxRate(double taxRate)
+    public ReturnsTaxRate TaxRate(decimal taxRate)
     {
         return TaxRate(taxRate.ToString());
     }
